refactor(runbotics-orchestrator-ui): :recycle: changed file structure of ProcessTile + refactor after second cr
