+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/Tile/ProcessTile/ProcessTile.styles.ts
@@ -1,46 +1,9 @@
 import {
-    CardContent,
-    SvgIcon,
     CardHeader,
     CardActionArea,
     Typography,
 } from '@mui/material';
 import styled from 'styled-components';
-
-export const StyledSvg = styled(SvgIcon)(({ theme }) => `
-    margin-left: ${theme.spacing(2)};
-    margin-right: ${theme.spacing(1)};
-`);
-
-export const StyledContent = styled(CardContent)`
-    display: grid;
-    grid-template-columns: 1fr 1.5fr;
-    grid-template-rows: repeat(2, auto);
-    justify-items: start;
-    justify-content: space-evenly;
-    width: 100%;
-
-    padding: 0 1.5rem 1rem 1.5rem;
-    gap: 1rem;
-`;
-
-export const Footer = styled.div`
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-
-    padding: 0.625rem 0.875rem;
-    gap: 2rem;
-
-    & > button {
-        padding: 0.3125rem;
-    }
-`;
-
-export const IconsWrapper = styled.div`
-    display: flex;
-    align-items: center;
-`;
 
 export const StyledCardHeader = styled(CardHeader)(({ theme }) => `
     &:hover {
