+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/Tile/ProcessTile/ProcessTileActions/index.ts
@@ -1,0 +1,3 @@
+import ProcessTileActions from './ProcessTileActions';
+
+export default ProcessTileActions;
