+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/Tile/ProcessTile/ProcessTileContent/index.ts
@@ -1,0 +1,3 @@
+import ProcessTileContent from './ProcessTileContent';
+
+export default ProcessTileContent;
