+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/Tile/ProcessTile/ProcessTileFooter/ProcessTileFooter.types.ts
@@ -1,5 +1,5 @@
 import { IProcess } from 'runbotics-common';
 
-export interface ProcessTileProps {
+export interface ProcessTileFooterProps {
     process: IProcess;
 }
