+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/Tile/ProcessTile/ProcessTileFooter/index.ts
@@ -1,0 +1,3 @@
+import ProcessTileFooter from './ProcessTileFooter';
+
+export default ProcessTileFooter;
