+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/Tile/ProcessTile/ProcessTileTag/ProcessTileTag.types.ts
@@ -1,0 +1,5 @@
+import { Tag } from 'runbotics-common';
+
+export interface ProcessTileTagProps {
+    tags: Tag[];
+}
