+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/Tile/ProcessTile/ProcessTileTag/index.ts
@@ -1,0 +1,3 @@
+import ProcessTileTag from './ProcessTileTag';
+
+export default ProcessTileTag;
