+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/Tile/ProcessTile/index.ts
@@ -1,10 +1,3 @@
 import ProcessTile from './ProcessTile';
 
 export default ProcessTile;
-export { default as ProcessTileActions } from './ProcessTile.actions';
-export { default as ProcessTileFooter } from './ProcessTile.footer';
-export { default as ProcessTileContent } from './ProcessTile.content';
-export { default as ProcessTileTags } from './ProcessTile.tags';
-export * from './ProcessTile.types';
-export * from './ProcessTile.styles';
-export * from './ProcessTile.utils';
